improve docs

+++ docs/deploy.py
@@ -3,22 +3,30 @@ import shutil
 import subprocess
 import sys
 
+
+def error(text):
+    print(f"\u001b[31;1m{text}\u001b[0m")
+    exit(1)
+
+
 cwd = os.getcwd()
 
 npm = shutil.which("npm")
 meson = shutil.which("meson")
+python = shutil.which("python")
 
 if npm is None:
-    print("Require npm")
-    exit(1)
+    error("Require npm")
 
-if npm is None:
-    print("Require meson")
-    exit(1)
+if meson is None:
+    error("Require meson")
+
+if python is None:
+    error("Require python")
 
 
 def step(name, args, cwd):
-    print("\u001b[32;1m-- " + name + " --\u001b[0m")
+    print(f"\u001b[32;1m-- {name} --\u001b[0m")
 
     code = subprocess.call(
         args=args,
@@ -27,19 +35,34 @@ def step(name, args, cwd):
     )
 
     if code != 0:
-        print("\u001b[31;1m-- Failure --\u001b[0m")
-        exit(1)
+        error("-- Failure --")
 
 
 def insert_gitignore(path):
+    print(f"\u001b[32;1m-- Insert gitignore: {path} --\u001b[0m")
     file = open(path + "/.gitignore", "w")
     file.write("*")
     file.close()
 
 
 def copy(src, dest):
-    os.makedirs(os.path.dirname(dest), exist_ok=True)
-    shutil.copy(src, dest)
+    print(f"\u001b[32;1m-- Copy from '{src}' to '{dest}' --\u001b[0m")
+    if os.path.isfile(src):
+        if len(os.path.dirname(dest)) > 0:
+            os.makedirs(os.path.dirname(dest), exist_ok=True)
+        shutil.copy(src, dest)
+    else:
+        shutil.copytree(src, dest)
+
+
+def remove(path):
+    print(f"\u001b[32;1m-- Remove '{path}' --\u001b[0m")
+    if not os.path.exists(path):
+        return
+    if os.path.isfile(path):
+        os.remove(path)
+    else:
+        shutil.rmtree(path)
 
 
 def prepare():
@@ -51,6 +74,9 @@ def prepare():
         "-Duselib_system=disabled",
         "-Duselib_bigint=enabled",
     ]
+
+    remove("site/src/generated")
+    remove("site/.vitepress/generated")
 
     step("Configure morphine",
          [meson, "setup", "docs/wasm/buildmorphine", "-Dprefix=" + cwd + "/wasm/deps", "-Ddefault_library=static",
@@ -68,21 +94,27 @@ def prepare():
     copy("wasm/out/bin/morphine.js", "site/.vitepress/generated/morphine.js")
     copy("wasm/out/bin/morphine.wasm", "site/.vitepress/generated/morphine.wasm")
 
-    # step("Generate docs", [gradle, "run", "--args", "../../components ../site/src/generated kt c"], "gen")
+    step("Generate markdown", [python, "extractor.py", "-p", "../components/**/*.[ch]", "-d", "site/src/generated"],
+         ".")
     copy("../changelog.md", "site/src/generated/changelog.md")
 
 
-if len(sys.argv) < 2:
-    variant = "dev"
-else:
-    variant = sys.argv[1]
+if __name__ == "__main__":
+    if len(sys.argv) < 2:
+        variant = "dev"
+    else:
+        variant = sys.argv[1]
 
-if variant == "dev":
-    prepare()
-    step("Deploy", [npm, "run", "docs:dev", '--'] + sys.argv[2:], "site")
-elif variant == "build":
-    prepare()
-    step("Build", [npm, "run", "docs:build"], "site")
-else:
-    print("\u001b[31;1mSupported variants: dev, build\u001b[0m")
-    exit(1)
+    if variant == "dev":
+        prepare()
+        step("Deploy", [npm, "run", "docs:dev", '--'] + sys.argv[2:], "site")
+    elif variant == "build":
+        prepare()
+        step("Build", [npm, "run", "docs:build"], "site")
+        remove("dist")
+        copy("site/.vitepress/dist", "dist")
+        insert_gitignore("dist")
+    elif variant == "clean":
+        step("Clean", [python, "cleaner.py"], ".")
+    else:
+        error("Supported variants: dev, build or clean")
